Next step: Review pheromones implementation.

+++ airplane.h
@@ -1,9 +1,9 @@
 
 struct airplane{
-	int appearance;
-	int earliest_lt;
-	int target_lt;
-	int latest_lt;
+	unsigned int appearance;
+	long int earliest_lt;
+	long int target_lt;
+	long int latest_lt;
 	float cost_before;
 	float cost_after;
 };
